+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 [full changelog](http://github.com/yolk/pump/compare/v0.6.2...master)
 
+* Pump::JSON allow partial output with :fields option
+
 ### 0.6.2 / 2013-07-26
 
 [full changelog](http://github.com/yolk/valvat/compare/v0.6.1...v0.6.2)